Fix TypeScript build errors for Vercel deployment
- Fix CreateCampaign.tsx: Change targetAmountInCents from BigInt to number (uint32)
- Fix FHEDonationHandler.tsx: Remove dead code and undefined variables
- Fix RealCampaignList.tsx: Use proper TypeScript syntax for window object access
- Clean up unused code and improve type safety

Build now passes successfully for Vercel deployment

+++ components/CreateCampaign.tsx
@@ -30,8 +30,8 @@ export default function CreateCampaign() {
     setIsLoading(true);
 
     try {
-      // Use target amount directly as uint256 (in USD cents for precision)
-      const targetAmountInCents = BigInt(parseInt(formData.targetAmount) * 100); // Convert USD to cents
+      // Use target amount as uint32 (in USD cents for precision)
+      const targetAmountInCents = parseInt(formData.targetAmount) * 100; // Convert USD to cents
       const durationInSeconds = parseInt(formData.duration) * 24 * 60 * 60; // Convert days to seconds
 
       await write({
@@ -41,7 +41,7 @@ export default function CreateCampaign() {
         args: [
           formData.name,
           formData.description,
-          targetAmountInCents, // Using cents for reasonable uint256 range
+          targetAmountInCents, // uint32 type for target amount
           BigInt(durationInSeconds),
         ],
       });
